some more changes.

+++ src/main/java/utils/chainCutter.java
@@ -1,9 +1,11 @@
 package utils;
+import blockchain.blockchain;
 import org.apache.commons.io.FileUtils;
 import proto.Types;
 import java.io.*;
 import java.nio.file.Path;
 import java.nio.file.Paths;
+import java.util.ArrayList;
 import java.util.List;
 
 public class chainCutter {
@@ -13,27 +15,33 @@ public class chainCutter {
         path = pathToDir;
         File dir = new File(pathToDir.toString());
         try {
-            FileUtils.forceDelete(dir);
+            if (dir.exists()) {
+                FileUtils.forceDelete(dir);
+            }
             FileUtils.forceMkdir(dir);
         } catch (IOException e) {
             logger.error("", e);
         }
 
     }
-    public static void cut(List<Types.Block> blocks) throws IOException {
-        for (Types.Block b : blocks) {
+    public static void cut(blockchain bc, int start, int end) throws IOException {
+        for (Types.Block b : new ArrayList<>(bc.getBlocksCopy(start, end))) {
             cutBlock(b);
+            bc.setBlock(b.getHeader().getHeight(), Types.Block.newBuilder()
+                    .setHeader(b.getHeader())
+                    .setTs(b.getTs())
+                    .build());
         }
     }
 
     static void cutBlock(Types.Block b) throws IOException {
+//        logger.info("Write block " + String.valueOf(b.getHeader().getHeight()));
         Path blockFile = Paths.get(path.toString(), String.valueOf(b.getHeader().getHeight()));
         File f = new File(blockFile.toString());
         f.createNewFile();
         try (FileOutputStream output = new FileOutputStream(blockFile.toString())) {
             b.writeDelimitedTo(output);
         }
-        b.getDataList().clear();
     }
 
     public static Types.Block getBlockFromFile(int height) throws IOException {
